Update README.md
This project was initiated to address the challenge of identifying underwater objects using sonar signals, a task crucial for maritime security and navigation. By employing machine learning algorithms, the project aims to provide an efficient and accurate solution for distinguishing between rocks and mines in sonar data. Through rigorous experimentation and analysis, the project seeks to contribute valuable insights into the field of underwater object detection and classification, with potential applications in naval operations, marine research, and environmental monitoring. Join us in our quest to develop robust models for rock vs mine prediction and make a positive impact in underwater exploration and security.